Combine the two StatusChartsManager extensions in Loop
